pyproject.toml
[project]
name = "geoglows_ecflow"
version = "3.1.1"
description = "ECFLOW RAPID workflow for GEOGLOWS"
authors = [
    { name = "Michael Souffront", email = "msouffront@aquaveo.com" },
    { name = "Riley Hales", email = "rchales@byu.edu" },
]
dependencies = [
    "pyyaml>=6.0.1",
    "numpy>=1.25.2",
    "netcdf4>=1.6.4",
    "dask>=2023.9.2",
    "pandas>=2.1.0",
    "xarray>=2023.8.0",
    "fastparquet>=2023.8.0",
    "zarr>=2.16.1",
    "boto3>=1.28.65",
    "basininflow>=0.13.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "BSD-3-Clause" }

[tool.coverage.report]
show_missing = true

[tool.setuptools.packages.find]
include = ["geoglows_ecflow*"]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project.scripts]
gdeploy = "geoglows_ecflow.bin.gdeploy:main"